app/requirements.txt
# Core dependencies
fastapi
uvicorn
Jinja2
pydantic-settings

# Database Dependencies
itsdangerous
asyncpg
sqlalchemy
psycopg

# Utility Dependencies
passlib
bcrypt
python-multipart

# Testing dependencies:
pytest 
pytest-asyncio 
httpx 
coverage

#for localization:
polib
